+++ feature/onboarding/src/main/java/com/twix/onboarding/contract/OnBoardingLoadingAction.kt
@@ -1,0 +1,8 @@
+package com.twix.onboarding.contract
+
+enum class OnBoardingLoadingAction {
+    CONNECT_COUPLE,
+    SUBMIT_PROFILE,
+    SUBMIT_DDAY,
+    SUBMIT_MARKETING_CONSENT,
+}
